% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gam.R
\name{estimate_gam}
\alias{estimate_gam}
\title{Estimate smooth changes in seropositivity and infection risk over time}
\usage{
estimate_gam(data_in)
}
\arguments{
\item{data_in}{A \verb{<data.frame>} containing individual-level serological data. The required columns
are "age" (in years) and "outcome" (0 or 1), representing a seropositive/seronegative result
according to the tested biomarker.}
}
\value{
A \verb{<data.frame>} with the probability of seropositivity by age and 95 CI, as well as annual proportion of the
population infected by year (where "age" represents the number of years into the past).
}
\description{
Fits a generalised additive model with monotone increasing SCOP-splines to binary
serological data and returns predictions about seropositivity by age and annual risk of historical infection
}
