% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_piecewise.R
\name{estimate_piecewise}
\alias{estimate_piecewise}
\title{Estimate step-wise changes in seropositivity and infection risk over time}
\usage{
estimate_piecewise(data_in, outbreak_years, survey_year)
}
\arguments{
\item{data_in}{A \verb{<data.frame>} containing individual-level serological data. The required columns
are "age" (in years) and "outcome" (0 or 1), representing a seropositive/seronegative result
according to the tested biomarker.}

\item{outbreak_years}{A \verb{<vector>} containing the years in which historical outbreaks occurred.}

\item{survey_year}{A \verb{<number>} giving the year in which the ages were reported.}
}
\value{
A \verb{<data.frame>} with the probability of seropositivity by age and 95 CI, as well as annual proportion of the
population infected by year (where "age" represents the number of years into the past).
}
\description{
Fits a generalised additive model with step-wise temporal risk (i.e. multiple outbreaks) to binary
serological data and returns predictions about seropositivity by age and annual risk of historical infection
}
