% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimate.R
\name{plot_estimate}
\alias{plot_estimate}
\title{Plot seropositivity and infection risk over time}
\usage{
plot_estimate(
  fit_in,
  outbreak_years = NULL,
  survey_year = NULL,
  write_file = NULL,
  y_range = c(0, 100)
)
}
\arguments{
\item{outbreak_years}{An optional \verb{<vector>} containing the years in which historical outbreaks occurred.}

\item{survey_year}{An optional \verb{<number>} giving the year in which the ages were reported.}

\item{write_file}{An optional \verb{<string>} giving the file path to write the plots to.}

\item{y_range}{An optional \verb{<vector>} giving alternative y axis range to plot infection risk on.}

\item{data_in}{A \verb{<data.frame>} containing the output of \code{estimate_gam} or \code{estimate_piecewise}}
}
\value{
A plot with the probability of seropositivity by age and 95 CI, as well as annual proportion of the
population infected by year.
}
\description{
Plots an output object from \code{estimate_gam} or \code{estimate_piecewise}
}
